Copper Expansion Part 1: Rings (#1920)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Copper is useless as a metal right now. It is only used as a precursor
to bronze and maybe for one specific niche recipe involving cogs. It is
literally trash and often the bane of miners who now have yet another
useless ore clumping up the spawn pool for iron, gems, gold and coal.

Except that this is about to change, and Tin is next. Copper can now be
used to make low-end jewelry -- Especifically, rings.

Copper rings can be combined with gems or be made as copper rings
themselves. Copper rings are 15 each, and Copper Gem rings have 73% of
the value of silver rings. Most of the value comes from the gem itself.

Those values are not arbitrary and are meant to make copper a worthwhile
material to make gem rings with. Decreasing the value further means that
it just is not worth it to waste gems on copper. I am willing to bump
down the value of copper rings themselves, though, or adjust the
stockpile prices.


![Capture](https://github.com/user-attachments/assets/488dda5d-b7b9-4a10-8245-8f8e758a9ede)



![Capture2](https://github.com/user-attachments/assets/bc1a113b-d072-4730-a21e-7a0fbd7d7d66)

(Includes on-mob ring sprite. You can see it on the right hand (His
right) of the in-game image if you zoom in. It's a single pixel.)

Also removes a duplicate silver ring entry.
Also gives a specific silver ring the correct base material.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

This PR opens the door for lower-end jewelry made of copper. Chokers and
other such things should come soon once I feel like it or someone beats
me to it.

Copper as is, is an useless ore that only serves as bloat. Visit any
mine in the game and you will see that miners don't bother collecting
this shit, and smiths don't bother working with it. This changes now.
Copper rings are going to be a work-intensive but powerful way for the
Steward to make money for the keep.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->