Increases damage dealt by bad repairs (#1940)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Changes damage from failed repairs to 5 brute to 10% of the objects
total damage. This makes spam repairing weapons cost a lot more to learn
smithing to level 1 but its still viable if you have a lot of practice
items you are willing to sacrifice.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Too many people are learning smithing by fucking up repairs
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->