Re-adjusts Tin and Copper ore price on Stockpile. (#1952)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Playing for a few rounds I realized that I probably made Copper too
pricey. Copper rings are not that expensive, and Artificer players were
having trouble getting the mammon to get copper for Bronze.

This PR makes both Tin and Copper a bit more expensive than iron, still
giving smiths a net-positive when used to make valuable items, but not
so expensive that other roles can not buy them at all.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Makes the balance of copper in the economy a bit more saner now that
I've had some experience seeing how its being used in play. This also
makes miners not get instantly rich when hitting veins of copper and
selling to the stockpile.

Meanwhile, this also makes roles that require copper like Architect and
Artificer be able to get them more easily. Copper's price was massively
overinflated for what it offered. Tin's was a bit too low with the new
Pewter ware.

Hopefully this will be the last price edit that I need to do for a
while.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->