Ports the Otavan mercs from Fablekeep ## About The Pull Request
The Otavan Routiers are a new unique noble merc class with all new custom armor sprites. This is a specialized class focused on defense, taking hits and protecting others. If Grenzelhoft is the offensive merc, Routiers are the defensive one. They also have noble blood. Their strength is relatively low compared to other mercs, but this is made up for with very high CON, leaning into the defensive purpose.
The corset on the armor may be toggled on and off via right click.

Not available to short races due to the lack of dwarf sprites for the gear.

Also adds a new sword called the Falchion, a new weapon wich deals less damage than a longsword but more defense

## Why It's Good For The Game

Cool looking new armor sprites along with new roleplay opportunities for mercenaries via having noble blood.

All the credits goes to @sylphynford, as this is only a porting of their contents with some descriptions changed. Tested locally and saw no issues, one can only hope..