Buffs Barkeep & Cook Jobs (#1924)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Gives the barkeep a satchel and cloth for cleaning, and the cook gets a
few extra item and skills when starting out to help them with their
roles.

## Why It's Good For The Game

The barkeep, who makes a lot of money, should have a satchel to store
things if they need to.

Additionally, the cook is now competitive with the cheesemaker towner
class (which is a glorified cook), but without the starting cheese and
less skills thank cheesemaker has still. The main skill changes being
fist fighting and wrestling to act as a bouncer in the kitchen if
needed. The cook also gets a little bit of farming like the cheesemaker
has, and also gets a satchel with a spare flint and cloth in it (for
cheese and cleaning like the barkeep). The cook also has a hunting knife
on their belt now too, which will help in making signs if they want to
make a menu board, or just assist in cutting things if they lose their
cleavers. Although the butcher knife still functions similarly to the
hunting knife, there are things it cannot do, and good cooks always have
a variety of knives at their disposal.

If any of these changes are too controversial, let me know.