Adding buyable shops (#1957)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
1)Added buyable shop, made so smith shop on the right would be bought
with coins.
Also nuked stairs to the flat above, want to test shops first then we
can add it back for town smiths depending on how people use place
following weeks.

![image](https://github.com/user-attachments/assets/52406c83-cf07-4555-959d-3a0881e8cce7)

![image](https://github.com/user-attachments/assets/80fb4cdb-afc0-4ac1-b898-c5aa928f778b)



2) Removed locks from graveyard since it rarely used

## Why It's Good For The Game

Want to test if people would actually use these shops, price for shops
is only 5 let's leave it at that it can be adjusted later. Main thing is
to test idea.
If not I will choose dif role for this road. Right now I want to make it
a shop street.