Rebalance prices around metal (#1961)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Rebalances the steward prices around worked iron costing 25 mammons to
transport and worked steel costing 35
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
This should make hiring a smith to make things cheaper instead of
instantly pressing a button and having it delivered. Hiring a skilled
specialist to do a job should be cheaper than having the equipment
imported over long distances while an apocalypse is happening
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->