Fixes Elvish Weapons Not Having Silver Effects (#1948)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Elvish weapons are made using silver. In spite of this, the Elvish Saber
and the Elvish Dagger were BOTH smelted into steel AND failed to apply
the silver effect that weapons made out of silver apply on werewolves
and vampires.

This PR fixes this and allows elvish sabers and elvish daggers to apply
the silver debuff to those antagonists.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

This PR improves item consistency, as those two weapons made out of
silver will now properly be smelted into silver.
Moreover, this also makes them powerful against Werewolves and Vampires,
as silver weapons should be.

I believe that not having those effects was an oversight when elf
weapons were added. Let me know if this is intentional, though.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->