Merchant rebalance (#1962)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Rebalances merchant costs to force a better reliance on smiths as well
as nerfing riddle and choker costs due to increasing tech level and
nerfing vault gameplay as well as giving freelance smiths an earlier
powerspike.
Reduces love potion costs as well.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Towners should be the number one source for cheaper gear as you have to
put more effort into interacting with them
Vault gaming is boring as fuck and being forced to buy riddles and
chokers to throw it in the vault is also boring.

Towner smiths generally dont get much business unless they are already
in a premade faction or keep due to the smiths guild having a great
forge, this leads to them not really being a viable playstyle even when
played round start as by the time you get the riddle everyone else has
already done business with the smiths guild resulting in a single player
gameplay loop.

Love potion costs were reduced so that more people get the option to use
it.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->