Adjusts prices to be in line with merchant price rebalances (#1968)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
See title
https://github.com/Rotwood-Vale/Ratwood-Keep/pull/1962/files#diff-5b3d70f579810d7332d4ab57e60d8fcae6d8663a2a6a338f7e363aab7169fc46
Lowers riddle and choker sellprices to 100(half of the buy price)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Oops
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->