Revert "Revert "Requested Sneak Changes & More"" (#1974)
Reverts Rotwood-Vale/Ratwood-Keep#1806

Deerjesus Completely lied.

>For context; they claimed the original PR was a stealth buff, which it
is; but only in specific instances. It makes sneaking more reliable vs
projectiles, for example.
>It is **however** NOT a buff specifically intended to JUST buff stealth
(it also buffs specific stealth playstyles)
>It IS HOWEVER, meant to **NERF** a ton of things, and change how
sneaking is used. This is intentional, it is meant to make sneaking
without the sneak skill HARDER, that was the POINT of the PR and it was
**INCLUDED IN BOLD TEXT**
>Their entire argument came to "well 3 people said they dont like it +
it goes against your PR design" (who are they to claim what I originally
intended in my PR???)

Unfixed two bugs (admins being unable to see sneaking people, walls
echoing incorrectly when people are sneaking).


If you want to 'fix' the PR, you can always tweak the multipliers that
control sneaking/etc. Nothing was wrong with how the code worked, just
how it was balanced...

Which is Gyran's fault for testmerging it with less than a day of
testing (but we don't talk about that)

frankly as nobody has noticed either change I'm genuinely considering if
sneaking just needs a full overhaul