Drakian Lore Addition (#1983)
## About The Pull Request
Adds pre-discussed lore to the Drakian species.
Cleared previously by Useroth, to my understanding. This was a long time
coming, only left excluded because we didn't have the extended
description to my understanding.

One minor issue is that I can't include an apostrophe for some reason in
the description, otherwise it breaks. So you have "ones" instead of
"one's". Sorry.

## Why It's Good For The Game
Loaded question. Next.