Fixes newmoon coat crafting recipe (#1986)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

The newmoon coat usually needs a Noc amulet, a rare-ish item,
considering how much more powerful they are to normal leather armor.
Someone forgot to put an "= 1" on the item list, making it NOT need said
amulet. This fixes it.

Photo of me testing it (It works, you can't craft it without an amulet
anymore)

![image](https://github.com/user-attachments/assets/c50327dd-6a55-4d17-aefa-a42289579da1)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Bug fix! And you can't make coats without the needed materials anymore.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->