Removes windows from the street smith shop (#1987)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Removes the side windows that only existed to cause pain, re adds the
funny pillar.

![image](https://github.com/user-attachments/assets/c3e0f3f0-7a48-4552-9960-148b65de736d)
Tiles in blue are the removed windows
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The funny pillar splits the shop in a nice way while the stone murder
holes make the interior unrealistically exposed even when the shutters
are closed.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->