Fix for Court Physician Dungeon Surgery (#1985)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
 
The Court Physician originally was not able to gain full access to the
patients waiting for him in the dungeon surgery room, neither did he
have the proper tools in order to perform surgery. With a small map
edit, I included additional forceps toward the back of the room(hard to
see with lack of lighting but it's there) and also changed the Cell Door
to a Holding Door which allows the Physician to now also have access to
the private cell room within dungeon surgery like he should have.

The only other solution here would be to give the Court Physician the
key:'keep_dungeon' in addition to 'keep_barracks' which he already has
but that may be slightly too advantageous and keeps the Court Physician
limited this way to his areas and not just open any cell, willy-nilly.

## Why It's Good For The Game

Fix for issue #953 

It gives full functionality back to the Court Physician if they
choose(and likely will) to use the dungeon surgery room.

![Screenshot 2024-12-19
135624](https://github.com/user-attachments/assets/420441a9-e2ac-42ab-9019-a7330deec095)