Oh noble wizard (#1993)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Makes court wizard noble
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
It is weird that the most powerfull wizard in whole dutchy, a person,
who cant be easily replaced and who works for duke himself, is
comsidered a commoner, while even councillors (who are definetly not so
important as mage) are noble.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->