Inquisition Pass #1
Modifies the skillset of Inquisitors and Confessors:
Provides the Inquisitor and Confessors with climbing and athletics 3, because they had neither of the skill prior.
Removes the sword and bow skill of the Confessor.
Provides Confessors with the same lockpick skill as the Inquisitor.

Modifies the items of Confessors:
Confessors now start with a lockpick, same as the Inquisitor.
Confessors given stun maces, to allow for non-lethal captures. Replaces the cudgel. Same damage as the cudgel, mind, when off, but six when on and a horrible rate of draining stamina for a stun. I don't know why I have to continually tell people this, but I intentionally made stun maces HORRIBLE and only effective against people who don't manage their stamina.

Misc for both:
Confessors and Inquisitors made soft-antags once again for objective tracking.
The round end report for those who've confessed has been modified. View the image to follow.

Inquisitor PQ requirement raised from 5 to 8.
Confessor PQ requirement raised from 0 to 4.

Confessor given the ability to torture folks for their antag status, same as the Inquisitor. As opposed to just their patron.

Torture is no longer impossible to do to people with masochism, because they can once again be hit with the torture debuff. Something impossible with a more recent change, which totally broke torture / faith testing on people with masochism as a flaw.