Small Surgery Buff (#1969)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Gives a small buff to surgery success chance so that being a physician
isn't a pain in the ass with constant step repeating from shitty rng

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Save surgeon players a headache and makes them slightly better and suck
slightly less compared to acolytes that get guaranteed success chances
and massive healing potential.