Inquisition Pass #1 (#2004)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Modifies the skillset of Inquisitors and Confessors: 
- Provides the Inquisitor and Confessors with climbing and athletics 3,
because they had neither of the skill prior.
 - Removes the sword and bow skill of the Confessor.
 - Provides Confessors with the same lockpick skill as the Inquisitor.

Modifies the items of Confessors:
 - Confessors now start with a lockpick, same as the Inquisitor.

Misc for both:
- Confessors and Inquisitors made soft-antags (Once again for
Confessors.) for objective tracking.
- The round end report for those who've confessed has been modified.
"The Damned", rather than "Confessors". View the image to follow.

![image](https://github.com/user-attachments/assets/901bfd39-5d1e-44a2-8dc0-40e81fb38bb0)

 - Inquisitor PQ requirement raised from 5 to 8.
 - Confessor PQ requirement raised from 0 to 4.
- Confessor given the ability to torture folks for their antag status,
same as the Inquisitor. As opposed to just their patron.
- Torture is no longer impossible to do to people with masochism,
because they can once again be hit with the torture debuff. Something
impossible with a more recent change, which totally broke torture /
faith testing on people with masochism as a flaw.

## Why It's Good For The Game
Much needed QoL changes. Stun mace for less-lethal captures. PQ
requirement was long overdue. Actual incentive to get confessions now.