Bookcases now work properly - You can put books back on the book shelf. (#2013)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

The origin of the bug is that the player could remove book cases from
the book case but not put them back. After reviewing the code for the
bookcases, I determined that there was never a method written to put the
books back. Now, I have included a method whereby if you touch the
bookcase with the book in hand, it will return that book to the
bookcase.

## Why It's Good For The Game

Fixes bug issue #1842 

This gives back functionality to the bookcase that it should have,
instead of refusing the user from putting books back. It even returns
the book to the list so that it can be grabbed by another player later
on if they wish.

Below are my screenshot tests for proof of it working.

![Screenshot 2024-12-21
150258](https://github.com/user-attachments/assets/0fa0adea-2a06-4cd4-8524-eff50714a4d1)
![Screenshot 2024-12-21
150306](https://github.com/user-attachments/assets/9478dc78-993f-461e-8671-ac096c69f594)
![Screenshot 2024-12-21
150316](https://github.com/user-attachments/assets/e03515b0-6cb2-4a5e-9f16-d4b112efb257)
![Screenshot 2024-12-21
150326](https://github.com/user-attachments/assets/4c4ae606-acd8-4dff-b388-1f482e12b4c8)
![Screenshot 2024-12-21
150337](https://github.com/user-attachments/assets/d13b8f74-9b9d-4b7e-8894-12e6a55821b1)