Gachas your fishing loot table + Adjusts fish sell price slightly (#2017)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Okay, this is pretty huge for fishing players and adds a bit more depth
and excitement to it so I'm going to explain what this PR does first.

Firstly, I adjusted the fish prices to be less than what they are.

Next, I added a ton of items to the fishing loot table which are mainly
junk, some useful, and some very valuable depending on if you are using
worms/grubs or leeches. The grub/worm loot tables are the same now, and
fisherman are able to now fish out junk from the waters. Leeches have an
extra set of items that they are exclusive to in order to keep the rare
things even more rare as leeches are harder to get.

Also bonus, I added shrimp to the loot table which was previously
unobtainable. It's as shrimple as that.

Here are the approximate rates for the loot tables:
Very Common: 25.0%
Common: 14.98%
Rare: 9.9%
Super Rare: 4.95%
Ultra Rare: 0.91% -- Leeches only.
Absolute Rarest: 0.5% -- Leeches only.

I know that should be accurate for leeches. It might be _slightly
higher_ rates for grubs/worms but they also have a reduced loot table.

## Why It's Good For The Game

Spent all day thinking and figuring this out so I'm going to try my best
to tell you why this is good.

For the price adjustments, currently fisherman just sit around all day
and just sell a shit load of fish and instantly become rich from the new
boat. I've seen longshoremen do that too, so now I hopefully fixed the
prices to be lowered and make it way less profitable in mass fishing to
earn money. You can still do it, and rare fish do sell good still but
less money now (I hope).

Now for the biggest change which is loot tables. The goal behind this is
to make fishing slightly more realistic and exciting by adding in
reasonable items that could get hooked on your line which people have
littered into the ocean or naturally comes ashore. The most common
things being sticks, fiber, applecores / general trash. Then you have
bits and pieces of clothing you can fish up too which fisherman can sell
to the seamster if they choose to keep their junk. It adds more depth to
fishing since now you can use your junk that you fish up to help
contribute to the economy. Lastly, if you are EXTREMELY lucky, you can
fish up rings, coin pouches, a single coin, etc. The leech table has
even better items but also incredibly rare, making fishing feel like a
past time players will want to try their luck at. Leeches are also
harder to dig up most of the time compared to the other two baits, I
think. This gives more player interaction with the merchant as a bonus
for players who do get lucky.

The current state of fishing is kinda sad and feels like it's been on
the back burner in my opinion, and this incentivizes more players to
fish and hang out. Players that don't specialize in fishing will also
take longer to get any chance at loot from fishing, which incentivizes
those serious enough at fishing to level up their skill to be able to
fish faster, as fishermen are woefully underplayed in the player base.
The aim of this PR is to add that 'luck of the sea' vibe and trying to
keep it balanced. I ran numbers through GPT several times so they should
be accurate for the ratios. I tested the PR locally as well to see how
well it worked and it was more fun than standard fishing.

If this PR goes through I would like to see how well the player base
gets the loot as it is a numbers game in the long run and although the
rates should be balanced on what drops, things do slip through the
cracks. Check the loot tables in the changelog for the full list of
items.

**Let me know your feedback. As someone who fishes a lot in the game, I
felt this was necessary to make it less boring, at least until we get
even more in-depth mechanics in the future.**