Bandit Smith | Hoardmaster Nerfening
Provides the bandits with a dedicated smith, in the role of the Forgehand.
Atop that, it significantly changes around the Hoardmaster by removing almost all class related objects, beyond the most basic required for their function. The rest can be smithed, as is the purpose of the Forgehand. The 'stuff' and 'things' categories have been completely untouched. Only class specific sections were touched, and this was, primarily, if not entirely, weapons and armour of a variety higher than the most basic for each respective class.

For example, the Knave can no longer purchase elven daggers, sabres or poisoned arrows, but he can still purchase iron daggers, swords and full quivers of bolts / arrows.
Special exceptions for classes are made, like the Sawbones, who can purchase bombs and killers ice still. Or the Mage, for example, who can buy scrying orbs and null magic rings.

Finally, it provides a basic smithery roundstart. The Forgehand, through the Hoardmaster, can purchase a riddle of steel for a proper forge, should he wish. (Why wouldn't he?)

Unrelated to the above, but it shifts around bandit packs into their own file path so they're not just laying about, and cleans up the innards just a tad of the files.