Canopies - You Forgot To Wash Them In Six Years And Now They Stink And Are Grimdark(tm) Update (#2030)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Makes the shop canopies grimmer and darker to better fit into the
environment. It also adds a blue canopy variant, which faces the side!

## Why It's Good For The Game

Useroth was not happy with the light and cheery canopies. By soaking
them in a mix of dirt, shit and piss, they now look just as shitty as
the rest of the town instead of sticking out as something from a
Nobledark Fantasy™. A side canopy now allows for a canopy to be added to
the sides of shops, rather than just vertically. They layer above
players correctly too, and add an extra variety of colour to the mix.
More variety!

BEFORE:

![b1](https://github.com/user-attachments/assets/cb4f31cc-2174-436a-80cc-352f3bde2f3f)
AFTER:

![a1](https://github.com/user-attachments/assets/b10e8594-33ad-48e9-8a8b-c322cf0e3dfc)

BEFORE:

![b2](https://github.com/user-attachments/assets/30ed096d-edac-43e9-8799-00eba26ed1c3)
AFTER:

![a2](https://github.com/user-attachments/assets/5467ab36-24ce-449f-bc8f-38ba2646ed2b)

BEFORE:

![b3](https://github.com/user-attachments/assets/f8bd6bce-5e01-412f-b8e0-bbd6827d142c)
AFTER:

![a3](https://github.com/user-attachments/assets/a445d0dc-23f1-4aaa-9585-dc968ae7cc12)

BEFORE:

![b4](https://github.com/user-attachments/assets/15c89e96-2e53-4aa2-a16a-b79974acaba4)
AFTER:

![a4](https://github.com/user-attachments/assets/2c559f13-fe4a-4917-9e9e-9b16721377b7)