PSYDON LIVES | Ecclesiast Changes (#2037)
_From all across Enigma they come, clad in purest silver and darkest
black. Holy men, fanatics, zealots of the Father of the Ten, their lives
devoted to the rooting of heresy, and the defense of Mankind from the
foul Inhumen threat, from the ravenous and cruel servants of the
Godhead, and ultimately, from themselves. Varied in race and creed, in
background and experience, there is one universal truth which binds
these men and women together to strike back against the shadow, which
gives them purpose in these darkest daes._

**_Psydon Lives._**

## About The Pull Request

This PR adds the ability for Psydonites to play the Ecclesiast Refugee
role, by making a few major changes to the subclasses in the role.

**Preacher -** When chosen as a Psydonite, the Preacher becomes,
mechanically, the Witch Hunter Adventurer, with some power scaling to
bring it closer in line with the rest of the Refugee rebalance. They
lose their ability to perform miracles or other Priest mechanics, but in
turn, are given the Inquisitor abilities to Force Confessions and
Torture.

**Returning Crusader** - When chosen as a Psydonite, the Returning
Crusader loses all Templar abilities. In return, they are given the holy
armor of the Silver Order of Psydon, a Steel Mace (and a 3 in Maces,
replacing their Sword skill), and a Lightweight Brigandine, as well as
Apprentice-level Treatment (So that they may heal themselves and/or
others without access to healing Miracles). They are also given a quick
bump to their Wrestling, to better wrangle (and inevitably crush) the
Deaddite slaves of Zizo.

**Monk** - Essentially unchanged, however, all Psydonites in the
Ecclesiast class now receive a Silver Psycross and a Puritan's Cloak to
identify themselves as Psydon's Faithful.

## Why It's Good For The Game

Psydon deserves some love, and the inability to play roaming Psydonites
(Monks, Witch Hunters, ect) due to the Refugee change negatively
effected the Psydonite players as a whole. This offers a bit more choice
and player agency when it comes to picking a Refugee role, while keeping
Ecclesiast at 5 slots to help stop a flood of Inquisitors or
blunt-focused Crusaders alongside the regular Ecclesiast classes.

Has been tested on local, and works entirely as intended.